Task 2/requirements/requirements.txt
transformers==4.49.0
tensorflow==2.18.0
pillow==11.1.0
numpy==2.0.2
scikit-learn==1.6.1